Initial code generation
Generated code for I2C peripheral to begin driver development. We may not go with I2C1 but it will do as a placeholder. If we change it it will be simple refactoring.

Configuration code for peripherals is in "Core/Inc" and "Core/Src" for .h and .c files specifically. .h files will need to be #include'd for any files or drivers that we write that need those peripherals.