Fix: Edit core.c Due to CAN Wrapper API Changes
Edit core.c so that it can compile. Note that a lot of functionality was
commented out as a temporary fix. The CAN Wrapper is moving to a new
repository, so even more changes will be coming. Once the new repo is
installed as a submodule, I will go back and fix the removed
functionality.